Refactor to set alias from python script
