+++ Dockerfile
@@ -269,7 +269,7 @@ ENV SHELL /bin/bash
 ADD systemuser.sh /srv/singleuser/systemuser.sh
 ADD userconfig.sh /srv/singleuser/userconfig.sh
 ADD userconfig.sh /srv/singleuser/userconfig.sh
-ADD configure_kernels.py /srv/singleuser/configure_kernels.py
+ADD configure_kernels_and_terminal.py /srv/singleuser/configure_kernels_and_terminal.py
 ADD executables/start_ipykernel.py /usr/local/bin/start_ipykernel.py
 RUN chmod 705 /usr/local/bin/start_ipykernel.py
